Quote tabele name in postgres
See #219 

I does not know Vala so I'm not sure this is the correct way to do it. It doesn't seem like Vala supports signle quoted strings wich otherwise would have been simpler.
Anyway even if this isn't correct, I guess it's clear what I'm trying to do so just close this PR and do it in the correct way instead.